fix: pass fork options directly to thread
