+++ lib/fork.js
@@ -71,7 +71,7 @@ const createWorker = (options, execArgv) => {
 			env: {NODE_ENV: 'test', ...process.env, ...options.environmentVariables, AVA_PATH},
 			execArgv,
 			workerData: {
-				cwd: options.projectDir
+				options
 			},
 			trackUnmanagedFds: true,
 			stdin: true,
