+++ lib/worker/base.js
@@ -1,13 +1,27 @@
 'use strict';
 const {pathToFileURL} = require('url');
+const path = require('path');
 const currentlyUnhandled = require('currently-unhandled')();
-require('./ensure-forked'); // eslint-disable-line import/no-unassigned-import
+const {isRunningInThread, isRunningInChildProcess} = require('./utils');
+
+// Check if the test is being run without AVA cli
+if (!isRunningInChildProcess && !isRunningInThread) {
+	const chalk = require('chalk'); // Use default Chalk instance.
+	if (process.argv[1]) {
+		const fp = path.relative('.', process.argv[1]);
+
+		console.log();
+		console.error(`Test files must be run with the AVA CLI:\n\n    ${chalk.grey.dim('$')} ${chalk.cyan('ava ' + fp)}\n`);
+
+		process.exit(1);
+	} else {
+		throw new Error('The ’ava’ module can only be imported in test files');
+	}
+}
 
 const channel = require('./channel');
 
-channel.send({type: 'ready-for-options'});
-
-channel.options.then(async options => {
+const run = async options => {
 	require('./options').set(options);
 	require('../chalk').set(options.chalkOptions);
 
@@ -32,7 +46,7 @@ channel.options.then(async options => {
 
 		dependencyTracking.flush();
 		await channel.flush();
-		process.exit(); // eslint-disable-line unicorn/no-process-exit
+		process.exit();
 	}
 
 	// TODO: Initialize providers here, then pass to lineNumberSelection() so they
@@ -227,11 +241,23 @@ channel.options.then(async options => {
 		channel.send({type: 'uncaught-exception', err: serializeError('Uncaught exception', true, error, runner.file)});
 		exit(1);
 	}
-}).catch(error => {
+};
+
+const onError = error => {
 	// There shouldn't be any errors, but if there are we may not have managed
 	// to bootstrap enough code to serialize them. Re-throw and let the process
 	// crash.
 	setImmediate(() => {
 		throw error;
 	});
-});
+};
+
+const noop = () => {};
+
+if (isRunningInThread) {
+	const {workerData} = require('worker_threads');
+	run(workerData.options).then(noop, onError); // eslint-disable-line promise/prefer-await-to-then
+} else if (isRunningInChildProcess) {
+	channel.send({type: 'ready-for-options'});
+	channel.options.then(run).catch(onError); // eslint-disable-line promise/prefer-await-to-then
+}
